% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{CopyButton}
\alias{CopyButton}
\title{Button for copying text to clipboard}
\usage{
CopyButton(
  id,
  label,
  text = "No Text",
  icon = icon("copy"),
  width = NULL,
  class = NULL,
  modal = FALSE
)
}
\arguments{
\item{id}{HTML ID to use for button.}

\item{label}{Text label of the button.}

\item{text}{Text to be copied in clipboard.}

\item{icon}{icon to be shown on the button before label.}

\item{width}{width of the button.}

\item{class}{CSS class for customizing button.}

\item{modal}{Whether button is used within bootstrap modal.}
}
\value{
Button
}
\description{
It copies text to clipboard using clipboardjs
}
\examples{
if (interactive()) {
 ui <- fluidPage(

  # Setup
  use_copy(),

  # Text Input 1
  textInput("text", "Enter Your Text"),

  # Copy Button 1
  CopyButton(
    "copybtn",
    "Copy to Clipboard",
    icon = icon("copy"),
    text = "No Text Found",
    class = "btn-primary",
    width = "150px"
  )

)
server <- function(input, output, session) {

  observe({

    req(input$copybtn)
    CopyButtonUpdate(session,
                     id = "copybtn",
                     label = "Copy to Clipboard",
                     icon = icon("copy"),
                     text = input$text
    )

  })

}

shinyApp(ui = ui, server = server)
}
}
\author{
Deepanshu Bhalla
}
