% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{CopyButtonUpdate}
\alias{CopyButtonUpdate}
\title{Update text to be copied on the client}
\usage{
CopyButtonUpdate(
  session,
  id = "copybtn",
  label = "Copy to Clipboard",
  icon = NULL,
  text = "Sample Text",
  modal = FALSE
)
}
\arguments{
\item{session}{Session object passed to function given to shinyServer}

\item{id}{HTML ID to use for button.}

\item{label}{Text label of the button.}

\item{icon}{icon to be shown on the button before label.}

\item{text}{Text to be copied in clipboard.}

\item{modal}{Whether button is used within bootstrap modal.}
}
\value{
Button
}
\description{
Update text to be copied in clipboard on the client
}
\examples{
if (interactive()) {
 ui <- fluidPage(

  # Setup
  use_copy(),

  # Text Input 1
  textInput("text", "Enter Your Text"),

  # Copy Button 1
  CopyButton(
    "copybtn",
    "Copy to Clipboard",
    icon = icon("copy"),
    text = "No Text Found",
    class = "btn-primary",
    width = "150px"
  )

)
server <- function(input, output, session) {

  observe({

    req(input$copybtn)
    CopyButtonUpdate(session,
                     id = "copybtn",
                     label = "Copy to Clipboard",
                     icon = icon("copy"),
                     text = input$text
    )

  })

}

shinyApp(ui = ui, server = server)
}
}
\author{
Deepanshu Bhalla
}
